requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
fastapi==0.73.0
h11==0.13.0
idna==3.3
pydantic==1.9.0
pymongo==4.0.1
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.4
